requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.4
packaging==24.0
pycparser==2.22
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
